refactor after vertex addition
