+++ setup/environment.yml
@@ -3,4 +3,4 @@ dependencies:
   - pip
   - python=3.8
   - pip:
-      - git+https://github.com/RedisVentures/redis-feast-gcp@main
+      - git+https://github.com/RedisVentures/redis-feast-gcp@vertex-triton-support
