+++ setup/setup.sh
@@ -7,6 +7,7 @@ gcloud auth activate-service-account $SERVICE_ACCOUNT_EMAIL \
 echo project_id = $PROJECT_ID > ~/.bigqueryrc
 
 # Enable APIs
+echo "\nEnabling GCP APIs"
 gcloud services enable artifactregistry.googleapis.com
 gcloud services enable ml.googleapis.com
 gcloud services enable aiplatform.googleapis.com
@@ -15,14 +16,16 @@ gcloud services enable cloudscheduler.googleapis.com
 gcloud services enable cloudfunctions.googleapis.com
 
 # Create Cloud Storage Bucket
+echo "\nCreating cloud storage bucket"
 gsutil ls -b gs://$BUCKET_NAME || gsutil mb gs://$BUCKET_NAME
 
 # Create BigQuery Dataset
-echo "Creating biqquery dataset"
+echo "\nCreating biqquery dataset"
 bq --location=us mk --dataset $PROJECT_ID:gcp_feast_demo
 
 # Deploy cloud Function
-echo "Creating cloud function for materialization"
+echo "\nCreating cloud function for materialization"
+mv setup/materialize.py .
 gcloud functions deploy feast-update-features \
     --source=./ \
     --entry-point=main \
@@ -36,7 +39,7 @@ gcloud functions deploy feast-update-features \
     --set-env-vars PROJECT_ID=$PROJECT_ID
 
 # Create Cloud Scheduler Job
-echo "Creating cloud scheduler task for triggering materialization daily"
+echo "\nCreating cloud scheduler task for triggering materialization daily"
 gcloud scheduler jobs create pubsub feast-daily-job \
     --location $GCP_REGION \
     --schedule "0 22 * * *" \
@@ -44,43 +47,69 @@ gcloud scheduler jobs create pubsub feast-daily-job \
     --message-body "This job schedules feature materialization once a day."
 
 # Create & Apply the Feature Store
-echo "\nCreating FEAST Feature Store"
+echo "\nCreating Feature Store"
 python setup/create.py
 python setup/apply.py
 
 # Setup Vertex AI and Triton
+echo "\nSetting up Triton Model Repository in GCS"
 CONTAINER_IMAGE_URI=$GCP_REGION-docker.pkg.dev/$PROJECT_ID/nvidia-triton/vertex-triton-inference
 ENDPOINT_NAME=vaccine-predictor-endpoint
 DEPLOYED_MODEL_NAME=vaccine-predictor
+MODEL_STORAGE_URI=gs://$BUCKET_NAME/models
 
 ## Upload Triton Model Repository Contents
 gsutil cp -r ./setup/models gs://$BUCKET_NAME/
-gsutil rm gs://$BUCKET_NAME/models/ensemble-model/1/.gitkeep
+gsutil rm $MODEL_STORAGE_URI/ensemble-model/1/.gitkeep
 
 ## Create Artifact Registry
+echo "\nCreating GCP Artifact Repository for Triton Serving Container"
 gcloud artifacts repositories create nvidia-triton \
   --repository-format=docker \
   --location=$GCP_REGION \
   --description="NVIDIA Triton Docker repository"
 
 ## Pull and Upload Triton Image
-NGC_TRITON_IMAGE_URI="ghcr.io/redisventures/triton-Python-FIL-server:22.08-py3"
+echo "\nPulling Triton Docker Image"
+NGC_TRITON_IMAGE_URI="ghcr.io/redisventures/tritonserver-python-fil:22.11-py3"
 docker pull $NGC_TRITON_IMAGE_URI
 docker tag $NGC_TRITON_IMAGE_URI $CONTAINER_IMAGE_URI
 
+echo "\nPushing Triton Docker Image to GCP"
 gcloud auth configure-docker $GCP_REGION-docker.pkg.dev
 docker push $CONTAINER_IMAGE_URI
 
-## Create Vertex AI Model
+# Create Vertex AI Model
+echo "\nCreating Vertex AI Model"
 gcloud ai models upload \
   --region=$GCP_REGION \
   --display-name=$DEPLOYED_MODEL_NAME \
   --container-image-uri=$CONTAINER_IMAGE_URI \
-  --artifact-uri=$REGISTRY_URI \
-  --container-args='--model-control-mode=poll --log-verbose=3 --log-info=1 --log-warning=1 --log-error=1'
+  --artifact-uri=$MODEL_STORAGE_URI \
+  --container-command="./entrypoint.sh" \
+  --container-env-vars="REDIS_CONNECTION_STRING=$REDIS_CONNECTION_STRING","REDIS_PASSWORD=$REDIS_PASSWORD","PROJECT_ID=$PROJECT_ID","GCP_REGION=$GCP_REGION","BUCKET_NAME=$BUCKET_NAME"
 
-## Create Endpoint
+# Create Endpoint
+echo "\nCreating Vertex AI Endpoint"
 gcloud ai endpoints create \
   --region=$GCP_REGION \
   --display-name=$ENDPOINT_NAME
 
+## Lookup Endpoint and Model IDs
+echo "\nDeploying Model to Endpoint"
+ENDPOINT_ID=$(gcloud ai endpoints list \
+  --region=$GCP_REGION \
+  --filter=display_name=$ENDPOINT_NAME \
+  --format="value(name)")
+
+MODEL_ID=$(gcloud ai models list \
+  --region=$GCP_REGION \
+  --filter=display_name=$DEPLOYED_MODEL_NAME \
+  --format="value(name)")
+
+# Deploy Model to the Endpoint on Vertex
+gcloud ai endpoints deploy-model $ENDPOINT_ID \
+  --region=$GCP_REGION \
+  --model=$MODEL_ID \
+  --display-name=$DEPLOYED_MODEL_NAME \
+  --machine-type=n1-standard-4
