--- train/requirements.txt
@@ -1,2 +1,0 @@
-xgboost==1.6.2
-scikit-learn
